requirements.txt
# Core Django
Django==4.2.9
django-ninja==1.1.0
djangorestframework==3.14.0  # Necesario para ValidationError y algunos serializers legacy
django-filter==23.5
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.9  # PostgreSQL
dj-database-url==2.1.0

# Authentication & Security
PyJWT==2.8.0
django-environ==0.11.2

# API Documentation
# Django Ninja incluye OpenAPI/Swagger automático
# DRF se mantiene solo para código legacy (ValidationError, serializers en infrastructure)

# Celery & Redis (for async tasks)
celery==5.3.4
redis==5.0.1

# Payment Providers
paypalrestsdk==1.13.1
stripe==8.0.0

# PDF Generation
reportlab==4.0.9
xhtml2pdf==0.2.13
WeasyPrint==60.2

# Email
django-ses==3.5.2  # AWS SES (optional)

# File Storage
boto3==1.34.25  # AWS S3
django-storages==1.14.2

# Utilities
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
python-decouple==3.8

# Monitoring & Logging
sentry-sdk==1.39.2

# Production Server
gunicorn==21.2.0
whitenoise==6.6.0

# Development tools (optional)
# django-debug-toolbar==4.2.0
# ipython==8.19.0

